requirements.txt
tensorflow
Pillow
matplotlib
opencv-python
numpy
pandas
